+++ buession-jdbc/src/main/java/com/buession/jdbc/datasource/config/AbstractPoolConfiguration.java
@@ -147,4 +147,5 @@ public abstract class AbstractPoolConfiguration implements PoolConfiguration {
 	public void setProperties(Properties properties){
 		this.properties = properties;
 	}
+
 }
